+++ package/api/routers/v1/despistes.router.mjs
@@ -1,0 +1,8 @@
+import express from 'express';
+import despistes from '../../controllers/despistes.controller.mjs';
+
+const router = express.Router();
+router.get('/', despistes.list);
+router.post('/', despistes.add);
+
+export default router;