+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Script to create the `clases` and `despistes` tables in a MySQL database.
+- Implementation of `GET` and `POST` methods in the API for managing `clases` and `despistes`.
 
 ## [0.1.0] - 2021-11-22
 ### Added
 - Flutter Project created for Android & iOS.
 
-[Unreleased]: https://github.com/BilalMoreno92/caza-despiste/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/BilalMoreno92/caza-despiste/compare/0.1.0...HEAD
+[0.1.0]: https://github.com/BilalMoreno92/caza-despiste/releases/tag/v0.1.0